# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/me/devices": {
      "get": {
        "operationId": "DevicesController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Active sessions across all platforms",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List active devices for the current user",
        "tags": [
          "Devices & Sessions"
        ]
      },
      "post": {
        "operationId": "DevicesController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register or update a device (upsert on device_id)",
        "tags": [
          "Devices & Sessions"
        ]
      }
    },
    "/v1/me/devices/pairing": {
      "post": {
        "operationId": "DevicesController_createPairing_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePairingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingTokenResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Issue a ten-minute, one-time mobile workspace pairing token",
        "tags": [
          "Devices & Sessions"
        ]
      }
    },
    "/v1/me/devices/pairing/consume": {
      "post": {
        "operationId": "DevicesController_consumePairing_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumePairingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairedWorkspaceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consume a pairing token after signing in on mobile",
        "tags": [
          "Devices & Sessions"
        ]
      }
    },
    "/v1/me/devices/current/{opaqueDeviceId}": {
      "delete": {
        "operationId": "DevicesController_revokeCurrent_v1",
        "parameters": [
          {
            "name": "opaqueDeviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke the caller device by its app-generated device_id",
        "tags": [
          "Devices & Sessions"
        ]
      }
    },
    "/v1/me/devices/{deviceId}": {
      "delete": {
        "operationId": "DevicesController_revoke_v1",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke a specific device / sign out of that session",
        "tags": [
          "Devices & Sessions"
        ]
      }
    },
    "/v1/me/devices/revoke-all": {
      "post": {
        "operationId": "DevicesController_revokeAll_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeAllDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sign out of all other devices",
        "tags": [
          "Devices & Sessions"
        ]
      }
    },
...
    "/v1/workspaces/{workspaceId}/import": {
      "get": {
        "operationId": "ImportsController_list_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportJobResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "List recent staged import jobs",
        "tags": [
          "Data portability"
        ]
      },
      "post": {
        "operationId": "ImportsController_create_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateImportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJobResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Parse, stage, and post a CSV, OFX, or QIF statement",
        "tags": [
          "Data portability"
        ]
      }
    },
    "/v1/workspaces/{workspaceId}/import/{jobId}": {
      "get": {
        "operationId": "ImportsController_get_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJobResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "tags": [
          "Data portability"
        ]
      }
    },
    "/v1/workspaces/{workspaceId}/export": {
      "get": {
        "operationId": "ImportsController_export_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Download the workspace ledger as CSV or PDF",
        "tags": [
          "Data portability"
        ]
      }
    }
  },
  "info": {
    "title": "NoorixFin API",
    "description": "Personal and household finance management API. All financial amounts are minor-unit decimal strings. All timestamps are ISO 8601 UTC. Every failure body matches the `ApiErrorBody` schema; branch on its `code`, never on `message`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "supabase-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase access token"
      }
    },
    "schemas": {
...
      "DeviceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "device_id": {
            "type": "string"
          },
          "platform": {
            "type": "string",
            "enum": [
              "web",
              "ios",
              "android"
            ]
          },
          "device_name": {
            "type": "string",
            "nullable": true
          },
          "os_version": {
            "type": "string",
            "nullable": true
          },
          "app_version": {
            "type": "string",
            "nullable": true
          },
          "last_seen_at": {
            "type": "string"
          },
          "last_ip": {
            "type": "string",
            "nullable": true
          },
          "first_seen_at": {
            "type": "string"
          },
          "revoked_at": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "device_id",
          "platform",
          "device_name",
          "os_version",
          "app_version",
          "last_seen_at",
          "last_ip",
          "first_seen_at",
          "revoked_at"
        ]
      },
      "RegisterDeviceDto": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "description": "Opaque app-generated UUID stored in SecureStore/sessionStorage"
          },
          "deviceName": {
            "type": "string",
            "description": "User-editable device label, e.g. \"Sharif's Pixel\""
          },
          "pushToken": {
            "type": "string",
            "description": "Push token or serialized Web Push subscription"
          },
          "pushProvider": {
            "type": "string",
            "enum": [
              "expo",
              "fcm",
              "apns",
              "webpush"
            ]
          }
        },
        "required": [
          "deviceId"
        ]
      },
      "CreatePairingDto": {
        "type": "object",
        "properties": {
          "workspaceId": {
            "type": "string"
          }
        },
        "required": [
          "workspaceId"
        ]
      },
      "PairingTokenResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "expires_at": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "expires_at"
        ]
      },
      "ConsumePairingDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "One-time token read from the pairing QR code"
          }
        },
        "required": [
          "token"
        ]
      },
      "PairedWorkspaceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "base_currency": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3
          }
        },
        "required": [
          "id",
          "name",
          "base_currency"
        ]
      },
      "RevokedResponseDto": {
        "type": "object",
        "properties": {
          "revoked": {
            "type": "boolean"
          }
        },
        "required": [
          "revoked"
        ]
      },
      "RevokeAllDto": {
        "type": "object",
        "properties": {
          "currentDeviceId": {
            "type": "string",
            "description": "Keep this device_id active (the caller's own device)"
          }
        }
      },
      "ImportRowResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "row_number": {
            "type": "number"
          },
          "normalized_payload": {
            "type": "object",
            "additionalProperties": true
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PROCESSING",
              "COMPLETED",
              "COMPLETED_WITH_ERRORS",
              "FAILED"
            ]
          },
          "error_message": {
            "type": "string",
            "nullable": true
          },
          "journal_entry_id": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "row_number",
          "normalized_payload",
          "status",
          "error_message"
        ]
      },
      "ImportJobResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "format": {
            "type": "string",
            "enum": [
              "CSV",
              "OFX",
              "QIF"
            ]
          },
          "filename": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PROCESSING",
              "COMPLETED",
              "COMPLETED_WITH_ERRORS",
              "FAILED"
            ]
          },
          "total_rows": {
            "type": "number"
          },
          "imported_rows": {
            "type": "number"
          },
          "failed_rows": {
            "type": "number"
          },
          "error_message": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          },
          "completed_at": {
            "type": "string",
            "nullable": true
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportRowResponseDto"
            }
          }
        },
        "required": [
          "id",
          "format",
          "filename",
          "status",
          "total_rows",
          "imported_rows",
          "failed_rows",
          "error_message",
          "created_at",
          "completed_at"
        ]
      },
      "CreateImportDto": {
        "type": "object",
        "properties": {
          "format": {
            "type": "string",
            "enum": [
              "CSV",
              "OFX",
              "QIF"
            ]
          },
          "content": {
            "type": "string",
            "description": "UTF-8 file content; maximum 5 MB"
          },
          "filename": {
            "type": "string",
            "example": "statement.csv"
          },
          "account_id": {
            "type": "string",
            "description": "Account affected by every imported statement row"
          },
          "expense_category_id": {
            "type": "string",
            "description": "Default category for negative/expense rows"
          },
          "income_category_id": {
            "type": "string",
            "description": "Default category for positive/income rows"
          },
          "idempotency_key": {
            "type": "string",
            "description": "A retry of this import returns the original job"
          }
        },
        "required": [
          "format",
          "content",
          "filename",
          "account_id",
          "expense_category_id",
          "idempotency_key"
        ]
      },
...
    }
  }
}